Add support for nonblocking operation suitable for event-driven frameworks
Functions amqp_recv_frames and amqp_pop_frame serve as a replacement for
amqp_simple_wait_frame. When the socket fd is set to nonblocking,
amqp_recv_frames will consume any data in the socket buffer until EAGAIN and
insert processed frames into the frame queue.

amqp_pop_frame can be called to consume frames off the queue.

This operation is incompatible with the existing blocking, synchronous rpc
calls. This is rectified through the addition of 'request' versions of method
calls which thinly wraps amqp_send_method and does not block on responses. This
places the responsibility on correlating responses with requests on the
higher-level api.

Signed-off-by: Duncan Burke <duncan.burke@orionvm.com.au>